DiadocClient fix AcquireCounteragent query params
