+++ src/diadoc/tests/client/conftest.py
@@ -2,6 +2,7 @@ from functools import partial
 import pytest
 
 from diadoc.client import DiadocClient
+from diadoc.models import DiadocPartner
 
 
 @pytest.fixture
@@ -13,3 +14,15 @@ def get_diadoc_fixture(get_fixture):
 def client(mocker):
     mocker.patch("diadoc.http.DiadocHTTP.token", return_value="TOKEN", new_callable=mocker.PropertyMock)
     return DiadocClient()
+
+
+@pytest.fixture
+def my_organization():
+    return DiadocPartner(
+        name="Пивзавод",
+        inn="771245768212",
+        kpp=None,
+        diadoc_id="75fcac12-ec63-4cdd-9076-87a8a2e6e8ba",
+        is_active=True,
+        is_roaming=False,
+    )
